✨(site) show the currently logged in user's email in the header
As the marsha site relies on a logged in user to function, we need
to display minimal information about it (eg. its email) in the
header so the actual user know they are logged in and which
account they are using as they use the marsha site.
